ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
boto3==1.34.0
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
